added underscore.js as a dependency (it's filter/find/select utility is sooooo nice).
added yarn.meds (MedicineCabiniet), provides a way of introducing "global/temporary" effects such as antibiotics into the game world.

implemented the bacterial side of the antibiotic unit-test (NEXT do the MedicineCabinet's side of the unit testing).

As part adding the ability to introduce "global/temp" effects for implementing antibiotics;
added onetime_update_callback to yarn.graph,
this method will be call the passed in closure against each GameObject in the game graph on the next available turn. At the end of the turn, the callback(s) are discarded.